% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_select_multiple_text.R
\name{delete_select_multiple_text}
\alias{delete_select_multiple_text}
\title{Delete select_multiple column text}
\usage{
delete_select_multiple_text(df, questionnaire)
}
\arguments{
\item{df}{The KoboToolbox dataset}

\item{questionnaire}{The "survey" sheet of your KoboToolbox survey}
}
\value{
The KoboToolbox dataset without the select_multiple columns
}
\description{
Delete select_multiple column text
}
\examples{
\dontrun{
survey_without_select_multiple <- delete_select_multiple_text(survey,
questionnaire = kobo_questionnaire)
}
}
