% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_multiple_select.R
\name{rename_multiple_select}
\alias{rename_multiple_select}
\title{Rename multiple_select columns}
\usage{
rename_multiple_select(
  df,
  variable_name,
  group_separator = "/",
  replacement = "."
)
}
\arguments{
\item{df}{The KoboToolbox dataset}

\item{variable_name}{The name of the variable that has multiple_select columns}

\item{group_separator}{The character (usually, "/" or "_") that separates the answer option and the group name}

\item{replacement}{The character that will replace the group_separator}
}
\value{
The KoboToolbox dataset with the multiple_select answer options of the corresponding variable renamed
}
\description{
Rename multiple_select columns
}
\examples{
\dontrun{
survey <- rename_multiple_select(survey,
variable_name = "type_food",
group_separator = "/",
replacement = ".")
}
}
