% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_meta_variables.R
\name{select_meta_variables}
\alias{select_meta_variables}
\title{Select meta variables}
\usage{
select_meta_variables(df, exceptions = "gps|uuid|submission_time")
}
\arguments{
\item{df}{The KoboToolbox dataset}

\item{exceptions}{A regular expression to exclude variables from the selection}
}
\value{
A character vector with the names of the meta variables
}
\description{
Select meta variables
}
\examples{
\dontrun{
meta_variables <- select_meta_variables(df = survey, exceptions = "uuid|submission_time")
}
}
