% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_relevant_variables.R
\name{select_relevant_variables}
\alias{select_relevant_variables}
\title{Select relevant variables from your survey}
\usage{
select_relevant_variables(
  questionnaire,
  variable_pattern =
    "^start$|^end$|today|select_one|select_multiple|integer|decimal|text"
)
}
\arguments{
\item{questionnaire}{The "survey" sheet of your KoboToolbox survey}

\item{variable_pattern}{A regular expression pattern to match the variable types you want to select}
}
\value{
A character vector of the names of the variables that match the pattern
}
\description{
Select relevant variables from your survey
}
\examples{
\dontrun{
relevant_survey_variables <- select_relevant_variables(questionnaire = kobo_questionnaire)
}
}
