% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_xml_to_label.R
\name{transform_xml_to_label}
\alias{transform_xml_to_label}
\title{Transform Kobo's XML codes to its corresponding labels}
\usage{
transform_xml_to_label(df, dictionary, exceptions)
}
\arguments{
\item{df}{The KoboToolbox dataset}

\item{dictionary}{The dictionary ("choices" sheet of your KoboToolbox questionnaire) that maps the XML codes to the labels}

\item{exceptions}{The columns that should not be transformed (usually the multiple_select columns)}
}
\value{
The KoboToolbox dataset with the XML codes transformed to the labels
}
\description{
Transform Kobo's XML codes to its corresponding labels
}
\examples{
\dontrun{
survey_with_labels <- transform_xml_to_label(survey,
dictionary = kobo_choices,
exceptions = c("type_food", "water_source"))
}
}
